hr_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-tenants>=3.5,<4.0
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.26
channels>=4.0
channels-redis>=4.1
django-redis>=5.4
celery>=5.3
redis>=5.0
requests>=2.31
reportlab>=4.0
kafka-python>=2.0.2
gunicorn>=21.2
django-environ>=0.11
django-cors-headers>=4.3
django-extensions>=3.2.3
python-decouple>=3.8
drf-yasg>=1.21,<2.0
